Implement strict query path matching
Previously, a query like "abc/foo" would match a path like "a/b/c/foo".
This is no longer the case; now, if a path separator is present in the
query, the matcher (by default) treats this as a hint that each "path
component" in the query must match a single component in the path being
matched (though matched path components don't need to be adjacent).

It's not yet clear if this behavior is most natural, so configurability
should be plumbed up to the Vim plugin in the future. It might also be
nice to add support for requiring that matched path components be
adjacent as well.