Matcher rewrite
The most user-visible aspect of the rewrite is a rewritten matching
algorithm. While optimizing for the same kinds of queries as before, and
using many of the same signals, the matcher now makes an attempt to
optimize parts of search items other than the basename. This in turn
allows us to improve basename matching behavior. Overall, matching
behavior should be roughly backwards-compatible, with considerably
better performance when matching queries with multiple path components.
(The only required change to the behavioral test is intentional.)