Quantize `crfile_basename_shared_prefix_len_` to words
`crfile_basename_shared_prefix_len_` is the part of a match's score that
represents the similarity between the currently open file's basename and
an item's. This is useful in matching header and source files, modules
and tests, etc. However, counting the number of characters results in a
long tail of spurious matches that happen to match the first character
or two. Count the number of matching words instead.