Use python-config script to detect Python flags
(aka "rewrite Python library finding, again")

Arch, and possibly others, places the Python .so outside of LIBPL;
there's a .a in LIBPL instead which FindPythonLibsCustom locates,
causing the resulting cpsm_py.so to be linked against the wrong
library (and also missing libutil, which causes `import cpsm_py`
to fail due to a missing `forkpty`). Use the python-config script
provided by the CPython distribution instead.

Fixes #35, #40